% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barcode_table}
\alias{barcode_table}
\title{Generate bar code table}
\usage{
barcode_table(
  lox,
  count_matrix = "all_samples",
  code_set = "all_codes",
  labels = "alias"
)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{count_matrix}{A matrix of the selected samples}

\item{labels}{specify if graph should be labelled by sample name or alias}

\item{samples}{list of samples by name}

\item{codes}{code_set selected}
}
\value{
A table containing the data about the bar codes
}
\description{
A function used to generate table for the barcode details.
}
