% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barcode_union}
\alias{barcode_union}
\title{Barcode union}
\usage{
barcode_union(rep1, rep2, type = "size", range)
}
\arguments{
\item{rep1}{loxcode object containing data of a sample}

\item{rep2}{loxcode object containing data of a sample}

\item{type}{string containing "size" or "complexity"}

\item{range}{A list containg the start and end of size or dist_origin to perform union on}
}
\value{
A unique list of the barcodes after performing union operation
}
\description{
Function to perform union of barcodes based on size or complexity
}
