% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bubble_plot}
\alias{bubble_plot}
\alias{bubble_plot,loxcode_experiment-method}
\title{Generate bubble from count_matrix}
\usage{
bubble_plot(
  loxcode_experiment,
  count_matrix = "all_samples",
  code_set = "all_codes",
  style = "ggplot",
  labels = "sample",
  clustering = "none",
  agglomeration = "complete",
  min_reads = 0,
  max_repeats = 100,
  min_repeats = 1,
  split_by1 = "none",
  split_by2 = "none"
)

\S4method{bubble_plot}{loxcode_experiment}(
  loxcode_experiment,
  count_matrix = "all_samples",
  code_set = "all_codes",
  style = "ggplot",
  labels = "sample",
  clustering = "none",
  agglomeration = "complete",
  min_reads = 0,
  max_repeats = 100,
  min_repeats = 1,
  split_by1 = "none",
  split_by2 = "none"
)
}
\arguments{
\item{loxcode_experiment}{loxcode_experiment object}

\item{count_matrix}{A matrix of the selected samples}

\item{code_set}{A matrix of the selected code sets}

\item{style}{string containing "ggplot" or "plotly"}

\item{labels}{A string containing the name by which to sample}

\item{clustering}{A string containing parameter by which clustering is to be performed}

\item{agglomeration}{Selection of the collection}

\item{min_reads}{Integer containing value for min number of reads}

\item{max_repeats}{Integer containing value for max number of repeats}

\item{min_repeats}{Integer containing value for min number of repeats}

\item{split_by1}{String containing parameter to perform first split by}

\item{split_by2}{String containing parameter to perform second split by}
}
\value{
Bubble plot of the selected samples. Also produced with splits and clusters in some cases.
}
\description{
Generate bubble plot from count_matrix and also perform clustering while plotting
}
\examples{
# Load required packages
library(LoxCodeR2024)

# Example usage
# Assume loxcode_experiment, count_matrix, and code_set are defined
loxcode_experiment <- readRDS("~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds")
bubble_plot(loxcode_experiment,count_matrix="all_samples",code_set="all_codes",style="ggplot", labels="sample", clustering="none",agglomeration="complete",min_reads=0,max_repeats=100,min_repeats=1,split_by1="none",split_by2="none")
}
