% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{code_frequency_pie}
\alias{code_frequency_pie}
\title{Code frequency pie}
\usage{
code_frequency_pie(x, s = "all_samples", c = "all_codes")
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{sample set of independent samples}

\item{c}{A matrix of the selected code sets}
}
\value{
scatterpie plot of the codes in every permutation of size and complexity
}
\description{
Display the frequency proportions of codes of every permutation of size and complexity
}
