% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{collapse_old}
\alias{collapse_old}
\alias{collapse_old,loxcode_experiment-method}
\title{Collapses the count matrix into samples with the sample metadata}
\usage{
collapse_old(x, count_matrix, collapse, name, union = TRUE, average = FALSE)

\S4method{collapse_old}{loxcode_experiment}(x, count_matrix, collapse, name, union = TRUE, average = FALSE)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{count_matrix}{current count_matrix}

\item{collapse}{column names of metadata on which to collapse}

\item{name}{name of new count_matrix}

\item{union}{boolean, True if barcodes in either should be counted}

\item{average}{boolean, True if counts should be averaged instead of summed}
}
\value{
loxcode_experiment object with new count_matrix
}
\description{
Collapses the count matrix into samples with the sample metadata
}
