% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{collapse_selection_old}
\alias{collapse_selection_old}
\alias{collapse_selection_old,loxcode_experiment-method}
\title{Collapse selected samples (old version)}
\usage{
collapse_selection_old(
  lox,
  s = "all_samples",
  c = "all_codes",
  i,
  n = NULL,
  union = TRUE,
  average = FALSE
)

\S4method{collapse_selection_old}{loxcode_experiment}(
  lox,
  s = "all_samples",
  c = "all_codes",
  i,
  n = NULL,
  union = TRUE,
  average = FALSE
)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{s}{sample set/ count_matrix}

\item{c}{code set}

\item{i}{indices of samples to collapse}

\item{n}{new sample name}

\item{union}{boolean: use union or intersection of samples}

\item{average}{boolean: use average of sum of counts}
}
\value{
new loxcode_experiment object
}
\description{
Collapse selected samples (old version)
}
