% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{correlation_plot}
\alias{correlation_plot}
\alias{correlation_plot,loxcode_experiment-method}
\title{correlation plot}
\usage{
correlation_plot(
  lox,
  count_matrix = "all_samples",
  code_set = "all_codes",
  method_ = "pearson"
)

\S4method{correlation_plot}{loxcode_experiment}(
  lox,
  count_matrix = "all_samples",
  code_set = "all_codes",
  method_ = "pearson"
)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{count_matrix}{A matrix of the selected samples}

\item{code_set}{A matrix of the selected code sets}

\item{method_}{correlation method (p)}
}
\description{
Function used to create a correlation plot
}
