% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decode}
\alias{decode}
\title{Decode FASTQ}
\usage{
decode(r, meta)
}
\arguments{
\item{r}{Paths of R1, 2 respectively}

\item{meta}{User-defined data-frame for sample metadata}
}
\value{
S4 loxcode_sample object with decoded results
}
\description{
Recover loxcodes from raw Illumina FASTQ output
}
