% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{delete_codeset}
\alias{delete_codeset}
\title{Deletes a specified codeset from a loxcode experiment}
\usage{
delete_codeset(x, n)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{n}{name of the codeset to be deleted}
}
\value{
new loxcode_experiment object with a new codeset
}
\description{
Deletes a specified codeset from a loxcode experiment
}
