% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{delete_count_matrix}
\alias{delete_count_matrix}
\title{Creates a new count_matrix with dataframe of samples}
\usage{
delete_count_matrix(x, c)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{c}{name of existing count_matrix to be deleted}
}
\value{
updated loxcode_experiment object
}
\description{
Creates a new count_matrix with dataframe of samples
}
