% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{density_plot}
\alias{density_plot}
\alias{density_plot,loxcode_sample-method}
\title{Density distributions}
\usage{
density_plot(
  ref,
  npois,
  dist_type = "sample",
  samples = list(),
  plot_type = "dist_orig"
)

\S4method{density_plot}{loxcode_sample}(
  ref,
  npois,
  dist_type = "sample",
  samples = list(),
  plot_type = "dist_orig"
)
}
\arguments{
\item{ref}{reference sample_distribution}

\item{npois}{poisson mean}

\item{dist_type}{sample distribution from "sample" or "poisson"}

\item{samples}{list of simulated samples}

\item{plot_type}{density plot of "size" or "dist_orig"}
}
\value{
density plot of the distribution of dist_orig or sizes
}
\description{
Plot the density of the distribution of dist_orig or sizes in simulated samples
}
