% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dist_count_beeswarm_plot}
\alias{dist_count_beeswarm_plot}
\title{Show recombination distance distribution by size as beeswarm plot}
\usage{
dist_count_beeswarm_plot(x, count_threshold)
}
\arguments{
\item{x}{loxcode_sample object}

\item{count_threshold}{counts threshold, barcodes with a count number exceeding this threshold are ignored. This is in order
to avoid having very large points resulting from barcodes with disproportionate read counts.}
}
\description{
Produces a beeswarm plot where for each cassette size, individual distinct barcodes are shown as
points. Size and color of points correspond to the read count of each barcode.
}
