% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dist_orig_plot}
\alias{dist_orig_plot}
\title{Plots distribution of dist_orig found in sample}
\usage{
dist_orig_plot(x, size)
}
\arguments{
\item{x}{loxcode_sample object}

\item{size}{loxcode size to consider}
}
\description{
Plots distribution of dist_orig found in sample
}
