% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dist_orig_plot_sample}
\alias{dist_orig_plot_sample}
\alias{dist_orig_plot_sample,loxcode_sample-method}
\title{Dist_orig Plot (v2)}
\usage{
dist_orig_plot_sample(sample)

\S4method{dist_orig_plot_sample}{loxcode_sample}(sample)
}
\arguments{
\item{sample}{loxcode sample object}
}
\value{
a plot of the distribution of distance from origin in the sample
}
\description{
Plots distribution of distance from origin found in sample. Works for loxcode_samples without loxcode_experiment
}
