% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{excise}
\alias{excise}
\alias{excise,lox_casette-method}
\title{Perform an excision on the sequence}
\usage{
excise(lox, a, b)

\S4method{excise}{lox_casette}(lox, a, b)
}
\arguments{
\item{lox}{lox_casette object}

\item{a}{lox_site position 1}

\item{b}{lox_site position 2}
}
\value{
an excised lox_casette object
}
\description{
Perform an excision on the sequence
}
\examples{
# Create a sample lox_casette object
lox <- create_lox_casette(...)

# Perform excision on the sequence
excise(lox, 10, 20)
}
