% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sets.R
\name{fillSetAliases}
\alias{fillSetAliases}
\title{fill set aliases
helper function complete the aliases for the specified sample set}
\usage{
fillSetAliases(lox, count_matrix)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{count_matrix}{count matrix of loxcode object}
}
\value{
loxcode object with filled alias
}
\description{
fill set aliases
helper function complete the aliases for the specified sample set
}
