% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{filtered_codes_pie}
\alias{filtered_codes_pie}
\alias{filtered_codes_pie,loxcode_experiment-method}
\title{Filtered code frequency pie}
\usage{
filtered_codes_pie(x, s = "all_samples", c = "all_codes", t = 0.05, m = 2)

\S4method{filtered_codes_pie}{loxcode_experiment}(x, s = "all_samples", c = "all_codes", t = 0.05, m = 2)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{independent sample set}

\item{c}{code set}

\item{t}{tolerance threshold for repeated proportions}

\item{m}{maximum repeats tolerated}
}
\value{
scatter pie of filtered codes
}
\description{
Filters out codes that have an overlap between independent samples
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assume x, s, c, t, and m are defined
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
filtered_codes_pie(lox, s = "all_samples", c = "all_codes", t = 5, m = 3)
}
