% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{filtered_codes_table}
\alias{filtered_codes_table}
\alias{filtered_codes_table,loxcode_experiment-method}
\title{Filtered Code freq table}
\usage{
filtered_codes_table(x, s = "all_samples", c = "all_codes", t = 0.05, m = 2)

\S4method{filtered_codes_table}{loxcode_experiment}(x, s = "all_samples", c = "all_codes", t = 0.05, m = 2)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{independent sample set}

\item{c}{A matrix of the selected code sets}

\item{t}{tolerance threshold for repeated proportions}

\item{m}{maximum repeats tolerated}
}
\value{
scatterpie of filtered codes
}
\description{
Data frame of filtered codes specified by user parameters
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assume x, s, c, t, and m are defined
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/Week2.rds")
filtered_codes_table(lox, s = "all_samples", c = "all_codes", t = 3, m = 5)
}
