% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gatherpairs}
\alias{gatherpairs}
\title{Gather Pairs}
\usage{
gatherpairs(
  data,
  ...,
  xkey = ".xkey",
  xvalue = ".xvalue",
  ykey = ".ykey",
  yvalue = ".yvalue",
  na.rm = FALSE,
  convert = FALSE,
  factor_key = FALSE
)
}
\arguments{
\item{data}{Data required for function}

\item{...}{variable as quosures}

\item{xkey}{.xkey}

\item{xvalue}{.xvalue}

\item{ykey}{.ykey}

\item{yvalue}{.yvalue}

\item{na.rm}{default as FALSE for remove na values}

\item{convert}{default as False}

\item{factor_key}{default as false}
}
\value{
A gathered pair of key pair values
}
\description{
Function used to
}
\examples{
# Load required packages
library(LoxCodeR2024)

# Example usage
# Assuming data is defined
gather_pairs(data)
}
