% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{generate_alias}
\alias{generate_alias}
\title{generate sample aliases}
\usage{
generate_alias(lox, s = "all_samples", meta)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{s}{sample set}

\item{meta}{columns of meta data on which to collapse}
}
\value{
loxcode_experiment object with aliases generated
}
\description{
generate sample aliases
}
