% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_alias}
\alias{get_alias}
\alias{get_alias,loxcode_experiment-method}
\title{Get sample alias}
\usage{
get_alias(lox, set, name)

\S4method{get_alias}{loxcode_experiment}(lox, set, name)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{set}{sample set name}

\item{name}{sample_name}
}
\value{
sample alias
}
\description{
Get sample alias
}
\examples{
# Load necessary libraries and data
library(LoxCodeR2024)
data <- read.csv("your_data.csv")

# Create a loxcode experiment object
lox_experiment <- create_lox_experiment(data)

# Get sample alias
alias <- get_alias(lox_experiment, "Sample_Set", "Sample_Name")
}
