% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_barcode_stats_rep}
\alias{get_barcode_stats_rep}
\title{Barcode Stats}
\usage{
get_barcode_stats_rep(union_bc, rep)
}
\arguments{
\item{union_bc}{Barcodes data to check in a sample}

\item{rep}{loxcode object containing data of a sample}
}
\value{
A object containing the barcode and data on the matched barcode in the sample
}
\description{
Number of times a barcode is seen in a sample
}
