% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_cass_vec}
\alias{get_cass_vec}
\title{Convert vector of cassette strings to list of numeric vectors}
\usage{
get_cass_vec(c)
}
\arguments{
\item{c}{vector of cassette strings}
}
\value{
list of numeric vector
}
\description{
Convert vector of cassette strings to list of numeric vectors
}
\examples{
# Example usage:
# Assuming 'cassette_strings' is a vector of cassette strings
# Convert the vector of cassette strings to a list of numeric vectors
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/Week2.rds")
# numeric_vectors <- get_cass_vec("test")
}
