% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{get_collapsed_meta2}
\alias{get_collapsed_meta2}
\alias{get_collapsed_meta2,loxcode_experiment-method}
\title{Get metadata of a merged experiment (old version)}
\usage{
get_collapsed_meta2(x, s)

\S4method{get_collapsed_meta2}{loxcode_experiment}(x, s)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{set of merged samples}
}
\value{
a data frame of meta data
}
\description{
Get metadata of a merged experiment (old version)
}
\examples{
# Load necessary libraries and data
library(LoxCodeR2024)
data <- read.csv("your_data.csv")

# Create a loxcode experiment object
lox_experiment <- create_lox_experiment(data)

# Merge samples
merged_samples <- merge_samples(lox_experiment, sample_set = "Sample_Set")

# Get collapsed metadata (old version)
meta <- get_collapsed_meta2(lox_experiment, merged_samples)
}
