% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_comparison_table}
\alias{get_comparison_table}
\title{Comparison Table}
\usage{
get_comparison_table(rep1, rep2, type = "size", range = c(1, 13))
}
\arguments{
\item{rep1}{loxcode object containing data of a sample}

\item{rep2}{loxcode object containing data of a sample}

\item{type}{string containing "size" or "complexity"}

\item{range}{A list containg the start and end of size or dist_origin to perform union on}
}
\value{
A matrix of comparision of two samples
}
\description{
Function used to compare two samples based on barcode counts
}
