% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_comparison_table2}
\alias{get_comparison_table2}
\title{Comparison Table}
\usage{
get_comparison_table2(s1, s2)
}
\arguments{
\item{s1}{loxcode object containing data of 1st sample}

\item{s2}{loxcode object containing data of 2nd sample}
}
\value{
A data frame of comparision of two samples
}
\description{
Function used to compare two samples based on barcode counts
}
\examples{
# Load required packages
library(LoxCodeR2024)

# Example usage
# Assuming s1 and s2 are defined
comparison_table(s1, s2)
}
