% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\docType{methods}
\name{get_data,loxcode_sample-method}
\alias{get_data,loxcode_sample-method}
\title{Access decoded cassette data}
\usage{
\S4method{get_data}{loxcode_sample}(x)
}
\description{
Access decoded cassette data
}
