% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{get_data}
\alias{get_data}
\title{Get unvalidated readout data for a given sample}
\usage{
get_data(x, s)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{sample name}
}
\value{
data.frame containing unvalidated readout data for sample s
}
\description{
Equivalent to performing loxcoder::data() on the sample s
}
