% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\docType{methods}
\name{get_origin_dist,loxcode_sample-method}
\alias{get_origin_dist,loxcode_sample-method}
\title{Fetch distances from origin}
\usage{
\S4method{get_origin_dist}{loxcode_sample}(x)
}
\description{
Fetch distances from origin
}
