% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\docType{methods}
\name{get_rec_prob,loxcode_sample-method}
\alias{get_rec_prob,loxcode_sample-method}
\title{Get recombination distance distribution}
\usage{
\S4method{get_rec_prob}{loxcode_sample}(x, size)
}
\description{
Get recombination distance distribution
}
