% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_samplename}
\alias{get_samplename}
\alias{get_samplename,loxcode_experiment-method}
\title{Get sample name from alias}
\usage{
get_samplename(lox, set, name)

\S4method{get_samplename}{loxcode_experiment}(lox, set, name)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{set}{sample set name}

\item{name}{alias}
}
\value{
sample alias
}
\description{
Get sample name from alias
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assuming lox, set, and name are defined
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/Week2.rds")
get_samplename(lox, set = "all_samples", name = "Sample 1")
}
