% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{get_valid}
\alias{get_valid}
\title{Get validated readout data for a given sample}
\usage{
get_valid(x, s)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{sample name}
}
\description{
Equivalent to performing loxcoder::valid() on the sample s
}
