% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stats_plot.R
\name{getBothPlot}
\alias{getBothPlot}
\title{get both plot
helper function to generate a bar graph of sizes/complexities for each sample}
\usage{
getBothPlot(countsTable, labels)
}
\arguments{
\item{countsTable}{table of counts by sample and size/complexity}

\item{labels}{label plot by sample_name or alias}

\item{plot}{plot type (both/ratio)}
}
\description{
get both plot
helper function to generate a bar graph of sizes/complexities for each sample
}
