% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getCountsPlot}
\alias{getCountsPlot}
\title{get counts plot
helper function to generate a bar graph of sizes/complexities for each sample}
\usage{
getCountsPlot(countsTable, labels, pos, data)
}
\arguments{
\item{countsTable}{table of counts by sample and size/complexity}

\item{labels}{label plot by sample_name or alias}

\item{pos}{fill or stack}

\item{data}{data to plot (size/complexity)}
}
\value{
ggplot object of plot
}
\description{
get counts plot
helper function to generate a bar graph of sizes/complexities for each sample
}
