% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stats_plot.R
\name{getCountsTable}
\alias{getCountsTable}
\title{get counts table
helper function to generate table of counts for each sample by selected data}
\usage{
getCountsTable(lox, samples, codes, count_matrix, counts, data)
}
\arguments{
\item{samples}{list of samples by name}

\item{codes}{code_set selected}

\item{count_matrix}{name of count_matrix selected}

\item{counts}{count_matrix selected}

\item{data}{data to count (size/complexity)}
}
\description{
get counts table
helper function to generate table of counts for each sample by selected data
}
