% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getData}
\alias{getData}
\title{Get sample data}
\usage{
getData(sample, plot_type = "size")
}
\arguments{
\item{sample}{loxcode sample object}

\item{plot_type}{"size" or "dist_orig"}
}
\value{
data frame with sizes in a sample
}
\description{
Get a list of sizes from the sample
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assuming sample is defined
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
getData(lox@samples$N712_N501)
}
