% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmap_plot}
\alias{heatmap_plot}
\alias{heatmap_plot,loxcode_experiment-method}
\title{Heatmap Plot}
\usage{
heatmap_plot(
  loxcode_experiment,
  count_matrix = "all_samples",
  code_set = "all_codes",
  style = "ggplot",
  labels = "sample",
  clustering = "none",
  agglomeration = "complete",
  min_reads = 0,
  max_repeats = 100,
  min_repeats = 1,
  split_by1 = "none",
  split_by2 = "none"
)

\S4method{heatmap_plot}{loxcode_experiment}(
  loxcode_experiment,
  count_matrix = "all_samples",
  code_set = "all_codes",
  style = "ggplot",
  labels = "sample",
  clustering = "none",
  agglomeration = "complete",
  min_reads = 0,
  max_repeats = 100,
  min_repeats = 1,
  split_by1 = "none",
  split_by2 = "none"
)
}
\arguments{
\item{loxcode_experiment}{loxcode_experiment object}

\item{count_matrix}{A matrix of the selected samples}

\item{code_set}{A matrix of the selected code sets}

\item{style}{string containing "ggplot" or "plotly"}

\item{labels}{A string containing the name by which to sample}

\item{clustering}{A string containing parameter by which clustering is to be performed}

\item{agglomeration}{setting default as complete. Used for clustering}

\item{min_reads}{Integer containing value for min number of reads}

\item{max_repeats}{Integer containing value for max number of repeats}

\item{min_repeats}{Integer containing value for min number of repeats}

\item{split_by1}{String containing parameter to perform first split by}

\item{split_by2}{String containing parameter to perform second split by}
}
\value{
Heat map plot of the selected samples. Also produced with splits and clusters in some cases.
}
\description{
Generate heatmap from count_matrix and also perform clustering while plotting
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assuming loxcode_experiment, count_matrix, and code_set are defined
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")
heatmap_plot(lox)
}
