% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\name{impute}
\alias{impute}
\alias{impute,loxcode_sample-method}
\title{Impute missing code in 13-element cassettes}
\usage{
impute(x)

\S4method{impute}{loxcode_sample}(x)
}
\arguments{
\item{x}{loxcode object}
}
\value{
cassette with imputed missing code
}
\description{
For 13-element cassettes that are missing a single element, the
missing element is imputed to minimise the resulting dist_orig.
}
\examples{
# Example usage:
# Assuming 'lox_data' is a loxcode object containing 13-element cassettes
# Impute the missing code in each 13-element cassette
lox <- readRDS("~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds")
#imputed_cassettes <- impute_missing_code_in_13_element_cassettes(lox_data)
#imputed_cassettes
}
