% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impute_13}
\alias{impute_13}
\title{Impute missing elements}
\usage{
impute_13(c, sizes)
}
\arguments{
\item{c}{list of loxcode numbers present in each cassete}

\item{sizes}{sizes of loxcode}
}
\description{
Impute missing elements
}
