% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impute_13_impl}
\alias{impute_13_impl}
\title{Impute missing element in a 13-element cassette}
\usage{
impute_13_impl(c2)
}
\arguments{
\item{c2}{string of number of lox present}
}
\value{
13-element cassette
}
\description{
Impute missing element in a 13-element cassette
}
\examples{
# Example usage:
# Assuming 'cassette_string' is a string representing the number of lox present
# Impute the missing element in the 13-element cassette
cassette <- impute_13_impl(cassette_string)
cassette
}
