% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{invert}
\alias{invert}
\title{Perform an inversion on the sequence}
\usage{
invert(lox, a, b)
}
\arguments{
\item{lox}{vector of lox sites}

\item{a}{site position 1}

\item{b}{site position 2}
}
\value{
a new vector of lox sites
}
\description{
Perform an inversion on the sequence
}
