% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_valid}
\alias{is_valid}
\title{Check if cassette is valid}
\usage{
is_valid(c)
}
\arguments{
\item{c}{a list of numeric vectors, or a character vector of decoded loxcodes}
}
\value{
TRUE if the cassette is valid, otherwise FALSE
}
\description{
Checks that cassette is an allowed size, even and odd elements are well-positioned
and that there are no repeated elements.
}
