% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\name{length,loxcode_sample-method}
\alias{length,loxcode_sample-method}
\title{Get number of rows in loxcode data}
\usage{
\S4method{length}{loxcode_sample}(x)
}
\arguments{
\item{x}{Loxcode object}
}
\value{
length of rows in loxcode
}
\description{
Get number of rows in loxcode data
}
