% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{load_from_xlsx}
\alias{load_from_xlsx}
\title{Create experiment}
\usage{
load_from_xlsx(name, s, dir, suffix_R1, suffix_R2, load = TRUE)
}
\arguments{
\item{name}{string, name of the experiment}

\item{s}{string, path to user-specified Excel file containing experiment samples.}

\item{dir}{string, path to directory containing fastq files}

\item{suffix_R1}{string, R1 suffix}

\item{suffix_R2}{string, R2 suffix}

\item{load}{boolean, whether to load samples or not (default is TRUE)}
}
\description{
Create a loxcode_experiment from xlsx sample table.
}
\details{
Must contain the following columns:

* `sample` -- sample name. This is used as the unique identifier for the sample

* `prefix` -- sample prefix. suffix_R1 and suffix_R2 are appended to this

* `meta` -- sample metadata. This is user specified

* `min_r1_len` -- minimum R1 length

* `min_r2_len` -- minimum R2 length
}
