% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\name{load_prob_files}
\alias{load_prob_files}
\title{Load probability tables (Markov chain)}
\usage{
load_prob_files(path)
}
\arguments{
\item{path}{path to directory containing the Markov chain tables}
}
\description{
Load probability tables (Markov chain)
}
