% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{class}
\name{lox_casette-class}
\alias{lox_casette-class}
\alias{lox_casette}
\title{Lox barcode casette object}
\value{
loxcode casette for simulation
}
\description{
Lox barcode casette object
}
\section{Slots}{

\describe{
\item{\code{lox_sites}}{vector of cre-binding sites}

\item{\code{code_elements}}{vector of distinguishable code elements}
}}

