% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\docType{class}
\name{loxcode_sample-class}
\alias{loxcode_sample-class}
\alias{loxcode_sample}
\title{S4 class to represent loxcode experimental sample data}
\description{
S4 class to represent loxcode experimental sample data
}
\section{Slots}{

\describe{
\item{\code{decode}}{A data.frame to contain raw decode data from loxcoder::decode()}

\item{\code{meta}}{A data.frame for user-defined sample metadata}
}}

