% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{make_codeset_index}
\alias{make_codeset_index}
\alias{make_codeset_index,loxcode_experiment-method}
\title{Creates a new codeset with dataframe of codesets testing}
\usage{
make_codeset_index(x, c, I, n)

\S4method{make_codeset_index}{loxcode_experiment}(x, c, I, n)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{c}{name of existing codeset to choose from}

\item{I}{indices to include}

\item{n}{name of the new codeset}
}
\value{
new loxcode_experiment object with a new codeset
}
\description{
Creates a new codeset with dataframe of codesets testing
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object
# Create a new codeset named 'new_codeset' by selecting indices 1, 3, and 5 from an existing codeset 'old_codeset'
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")
# new_experiment <- make_codeset_index(experiment, c = 'old_codeset', I = c(1, 3, 5), n = 'new_codeset')
# new_experiment
}
