% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sets.R
\name{make_count_matrix}
\alias{make_count_matrix}
\alias{make_count_matrix,loxcode_experiment-method}
\title{Creates a new count_matrix with dataframe of samples}
\usage{
make_count_matrix(lox, count_matrix = "all_samples", indices, name = "test")

\S4method{make_count_matrix}{loxcode_experiment}(lox, count_matrix = "all_samples", indices, name = "test")
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{count_matrix}{name of existing count_matrix to choose from}

\item{indices}{indices to include}

\item{name}{name of the new count_matrix}
}
\value{
new loxcode_experiment object with a new count_matrix
}
\description{
Creates a new count_matrix with dataframe of samples
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object
lox <- readRDS("~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds")
# Create a new count_matrix named 'new_count_matrix' by selecting indices 1, 3, and 5 from an existing count_matrix 'old_count_matrix'
new_experiment <- make_count_matrix(lox,"all_samples",c(1,3,5))
new_experiment
}
