% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{make_filtered_codeset}
\alias{make_filtered_codeset}
\title{Create new filtered code set according to defined parameters}
\usage{
make_filtered_codeset(x, s, c, t, m, n)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{independent sample set}

\item{c}{code set}

\item{t}{tolerance threshold for repeated proportions}

\item{m}{maximum repeats tolerated}

\item{n}{new code set name}
}
\value{
new loxcode_experiment object
}
\description{
Create new filtered code set according to defined parameters
}
