% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\docType{methods}
\name{makeid,loxcode_sample-method}
\alias{makeid,loxcode_sample-method}
\title{Append a colum of packed cassette IDs, or -1 if it cannot be packed}
\usage{
\S4method{makeid}{loxcode_sample}(x)
}
\description{
Append a colum of packed cassette IDs, or -1 if it cannot be packed
}
