% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{merge_by}
\alias{merge_by}
\title{Merge samples by label}
\usage{
merge_by(x, by)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{by}{name of column in samp_table to merge by}
}
\value{
loxcode_experiment object containing merged samples
}
\description{
Merge samples present in a loxcode_experiment object according to a specified label column in
samp_table. All samples with a common value for the label column are merged together. The resulting
loxcode_experiment object contains loxcode_samples named by the corresponding labels in by.
}
