% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{merge_experiments2}
\alias{merge_experiments2}
\title{Merge a vector of experiments}
\usage{
merge_experiments2(experiments, name)
}
\arguments{
\item{experiments}{vector of loxcode_experiment objects}

\item{name}{name of new loxcode_experiment object}
}
\value{
loxcode_experiment object
}
\description{
Merge a vector of experiments
}
