% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{merge_sample}
\alias{merge_sample}
\title{Merge two loxcode_sample objects}
\usage{
merge_sample(s1, s2)
}
\arguments{
\item{s1}{sample 1}

\item{s2}{sample 2}
}
\value{
loxcode_sample object corresponding to merged sample
}
\description{
Barcodes from s1 and s2 are merged. For barcodes present in both samples, read counts are
summed. Sample descriptors and other members are concatenated in order.
}
\examples{
# Example usage:
# Assuming 'sample1' and 'sample2' are loxcode_sample objects
# Merge the two samples into a new loxcode_sample object
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")
merged_sample <- merge_sample(lox@samples$N712_N501,lox@samples$N712_N506)
merged_sample
}
