% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{min_flip_dist}
\alias{min_flip_dist}
\title{Distance at which the flip happened}
\usage{
min_flip_dist(c, size, v)
}
\arguments{
\item{c}{list of loxcode numbers present in each cassette}

\item{size}{sizes of loxcode}

\item{v}{flip distance}
}
\value{
Distance at which the flip happened
}
\description{
Distance at which the flip happened
}
\examples{
# Example usage:
# Assuming 'cassette_list' is a list of loxcode numbers present in each cassette,
# 'loxcode_sizes' is a vector containing the sizes of the loxcode, and 'flip_distance_value' is the flip distance.
# Calculate the distance at which the flip happened
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/Week2.rds")
# flip_dist <- min_flip_dist(cassette_list, loxcode_sizes, flip_distance_value)
# flip_dist
}
