% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R, R/loxcode_sample.R
\name{name<-}
\alias{name<-}
\title{Set experiment name}
\usage{
name(x, v) <- value

name(x, v) <- value
}
\description{
Set experiment name

Set loxcode sample name (description)
}
