% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R, R/loxcode_sample.R
\name{name}
\alias{name}
\title{Fetch experiment name}
\usage{
name(x)

name(x)
}
\arguments{
\item{x}{loxcode_experiment object}
}
\value{
name of experiment
}
\description{
Fetch experiment name

Get loxcode sample name (description)
}
