% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{pair_comparison_plot,loxcode_sample-method}
\alias{pair_comparison_plot,loxcode_sample-method}
\title{Plots pair comparison of two loxcode_sample objects}
\usage{
\S4method{pair_comparison_plot}{loxcode_sample}(x)
}
\description{
Plots pair comparison of two loxcode_sample objects
}
