% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pair_comparison_plot}
\alias{pair_comparison_plot}
\title{Produce pair comparison scatterplot}
\usage{
pair_comparison_plot(x1, x2, dist_range, ...)
}
\arguments{
\item{x1}{loxcode_sample object for sample 1}

\item{x2}{loxcode_sample object for sample 2}

\item{dist_range}{c(lower, upper), specifying the range of dist_orig values to consider. Both bounds are inclusive.}
}
\description{
For each barcode in x1, x2, read counts in each sample are plotted on a log-scale, after adjusting sample x2 to match the
total read count (for valid barcodes) in sample x1. Barcodes present in one sample and not the other are shown
horizontally and vertically (below 0) using beeswarm plots
}
