% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pair_comparison_plot2}
\alias{pair_comparison_plot2}
\alias{pair_comparison_plot2,loxcode_experiment-method}
\title{Pair Comparison Plot 2}
\usage{
pair_comparison_plot2(
  lox,
  sampleset = "all_samples",
  codeset = "all_codes",
  s1,
  s2,
  colorBy = "size",
  labels = "alias",
  sizeRange = NULL,
  dist_origRange = NULL,
  firstreadRange = NULL
)

\S4method{pair_comparison_plot2}{loxcode_experiment}(
  lox,
  sampleset = "all_samples",
  codeset = "all_codes",
  s1,
  s2,
  colorBy = "size",
  labels = "alias",
  sizeRange = NULL,
  dist_origRange = NULL,
  firstreadRange = NULL
)
}
\arguments{
\item{lox}{Loxcode experiment object}

\item{sampleset}{A matrix of the selected samples}

\item{codeset}{A matrix of the selected code sets}

\item{s1}{First Loxcode sample}

\item{s2}{Second Loxcode sample}

\item{colorBy}{Color points by size or dist_orig or firstread}

\item{labels}{Sample name or alias as labels}

\item{sizeRange}{Range of size available}

\item{dist_origRange}{Range of distance from origin available}

\item{firstreadRange}{Range of first read available}
}
\value{
A plot containing pairwise comparison of the samples
}
\description{
Function to create a pair comparision plot by passing code set and sample set. This function can use custom sample and code set for plot creation.
}
\examples{
# Load required packages
library(LoxCodeR2024)

# Example usage
# Assuming lox, sampleset, codeset, s1, and s2 are defined
pair_comparison_plot2(lox, sampleset, codeset, s1, s2, colorBy = "size", labels = "Sample Name", sizeRange = c(10, 100), dist_origRange = c(0, 100), firstreadRange = c(0, 100))
}
