% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pair_comparison_plot_all}
\alias{pair_comparison_plot_all}
\title{Pair Comparison Plot all}
\usage{
pair_comparison_plot_all(
  lox,
  min_reads = 10,
  matrix = "all_samples",
  codeset = "all_codes"
)
}
\arguments{
\item{lox}{Loxcode_experiment object}

\item{min_reads}{Code set}

\item{matrix}{First Loxcode_sample}

\item{codeset}{Second Loxcode_sample}
}
\value{
A pair comparision plot comparing all samples agaist each other
}
\description{
Pair Comparison Plot all
}
