% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stats_plot.R
\name{readstats_plot}
\alias{readstats_plot}
\title{readstats plot (v3)}
\usage{
readstats_plot(
  lox,
  count_matrix = "all_samples",
  code_set = "all_codes",
  plot = "size",
  fill = TRUE,
  labels = "alias"
)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{plot}{specify which plot (size/complexity/both/ratio)}

\item{fill}{specify whether reads should be normalized in boxplots}

\item{labels}{specify if graph should be labelled by sample name or alias}
}
\value{
Plot based on the type of plot selected
}
\description{
Function to create size, complexity, both or ratio plot
}
