% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{readstats_plot2}
\alias{readstats_plot2}
\title{readstats plot}
\usage{
readstats_plot2(
  loxcode_experiment,
  count_matrix = "all_samples",
  code_set = "all_codes",
  plot = "size",
  fill = TRUE,
  labels = "alias"
)
}
\arguments{
\item{loxcode_experiment}{loxcode_experiment object}

\item{count_matrix}{A matrix of the selected samples}

\item{code_set}{A matrix of the selected code sets}

\item{plot}{specify which plot (size/complexity/both/ratio)}

\item{fill}{specify whether reads should be normalized in boxplots}
}
\value{
A plot of the selected type explaing details of the samples
}
\description{
Generate size, complexity, both, ratio plot for the samples
}
