% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{rename_codeset}
\alias{rename_codeset}
\alias{rename_codeset,loxcode_experiment-method}
\title{rename code set}
\usage{
rename_codeset(x, c, n)

\S4method{rename_codeset}{loxcode_experiment}(x, c, n)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{c}{code set}

\item{n}{new code set name}
}
\value{
new loxcode_experiment object
}
\description{
rename code set
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object
# Rename the code set 'old_code_set' to 'new_code_set'
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")
new_experiment <- rename_codeset(lox,'all_codes','test')
new_experiment
}
