% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{rename_sample}
\alias{rename_sample}
\alias{rename_sample,loxcode_experiment-method}
\title{Renames a sample in a loxcode_experiment}
\usage{
rename_sample(x, c, n, o)

\S4method{rename_sample}{loxcode_experiment}(x, c, n, o)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{c}{count_matrix name}

\item{n}{new sample name}

\item{o}{old sample name}
}
\value{
new loxcode_experiment object with updated sample name
}
\description{
Renames a sample in a loxcode_experiment
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object and 'old_name' is the old sample name to be renamed
# Rename the sample named 'old_name' to 'new_name' in the count matrix 'count_matrix_name'
lox <- readRDS('~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds')
#new_experiment <- rename_sample(x = experiment, c = 'count_matrix_name', n = 'new_name', o = 'old_name')
#new_experiment
}
