% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sets.R
\name{rename_sampleset}
\alias{rename_sampleset}
\alias{rename_sampleset,loxcode_experiment-method}
\title{rename sample set}
\usage{
rename_sampleset(x, s, n)

\S4method{rename_sampleset}{loxcode_experiment}(x, s, n)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{sample set}

\item{n}{new code set name}
}
\value{
new loxcode_experiment object
}
\description{
rename sample set
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object and 'old_set_name' is the old sample set name to be renamed
# Rename the sample set named 'old_set_name' to 'new_set_name'
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
new_experiment <- rename_sampleset(lox,"all_samples","new")
new_experiment
}
