% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{retrieve_dist_origin}
\alias{retrieve_dist_origin}
\title{Assumes that input cassette is valid already!}
\usage{
retrieve_dist_origin(c, sizes)
}
\arguments{
\item{c}{input cassette}

\item{sizes}{sizes of all loxcode}
}
\description{
Assumes that input cassette is valid already!
}
