% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{retrieve_dist_origin_single}
\alias{retrieve_dist_origin_single}
\title{Assumes that input cassette is valid already!}
\usage{
retrieve_dist_origin_single(c, size)
}
\description{
Assumes that input cassette is valid already!
}
