% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{retrieve_dist_pair}
\alias{retrieve_dist_pair}
\title{Assumes that input cassette is valid already!}
\usage{
retrieve_dist_pair(c1, c2)
}
\arguments{
\item{c1}{input cassette 1}

\item{c2}{input cassette 2}
}
\value{
distance between 2 cassette
}
\description{
Assumes that input cassette is valid already!
}
\examples{
# Example usage:
# Assuming 'cassette1' and 'cassette2' are valid input cassettes.
# Calculate the distance between the two cassettes
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/Week2.rds")
#dist_between_cassettes <- retrieve_dist_pair(cassette1, cassette2)
#dist_between_cassettes
}
