% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\docType{methods}
\name{retrieve_prob_ensemble,loxcode_sample-method}
\alias{retrieve_prob_ensemble,loxcode_sample-method}
\title{Get generation probability using ensemble}
\usage{
\S4method{retrieve_prob_ensemble}{loxcode_sample}(x)
}
\description{
Get generation probability using ensemble
}
