% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_markov}
\alias{run_markov}
\title{Calculate generation probabilities using Markov chain formulation}
\usage{
run_markov(n, T)
}
\arguments{
\item{n}{number of steps to generate}

\item{T}{transition probabilities specified as interaction probabilities}
}
\description{
T specifies the transition probabilities by assigning probabilities to the interactions 
between LoxP sites. T must be a list of length 5 where each entry is a NumericMatrix.
T[[1]] must be a 14x14 matrix where element (i, j) = (j, i) corresponds to the probability 
of interaction between LoxP sites i and j, where LoxP sites are numbered 1-14 (or 0-13) inclusive.
For example, starting from the unrecombined cassette, an interaction (0, 12) yields the 1-element cassette {13}
}
\details{
Calculates the generation probability distribution for cassettes
for varying values of n (step number)
}
