% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{sample}
\alias{sample}
\title{Get loxcode_sample object}
\usage{
sample(x, s)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{string, sample name}
}
\description{
Get loxcode_sample object
}
