% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sample_comparison_pie}
\alias{sample_comparison_pie}
\title{sample comparison pie}
\usage{
sample_comparison_pie(
  x,
  count_matrix = "all_samples",
  code_set = "all_codes",
  scale = 1,
  labels = "sample"
)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{count_matrix}{A matrix of the selected samples}

\item{code_set}{A matrix of the selected code sets}

\item{scale}{scale the size of pies}

\item{labels}{by sample or alias}
}
\value{
A plot containg comparision of the samples in the form of a pie
}
\description{
Pie comparion plot of the samples
}
