% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{sampnames}
\alias{sampnames}
\title{Fetch names of samples}
\usage{
sampnames(x)
}
\arguments{
\item{x}{loxcode_experiment object}
}
\value{
names (handles) of samples
}
\description{
The names returned by this function can be used to refer to each individual sample
}
