% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{saturation_multi}
\alias{saturation_multi}
\title{Multiple lined saturation plot}
\usage{
saturation_multi(
  loxcode_experiment,
  loxcode_samples,
  codesets,
  colorby = "dist_orig"
)
}
\arguments{
\item{loxcode_experiment}{loxcode_experiment object}

\item{loxcode_samples}{A matrix of the selected samples}

\item{colorby}{colour by dist_orig or sample}

\item{code_set}{A matrix of the selected code sets}
}
\value{
Saturation plot of the loxcode sample
}
\description{
Multiple lined saturation plot
}
