% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{saturation_plot}
\alias{saturation_plot}
\title{Saturation plot}
\usage{
saturation_plot(loxcode_experiment, loxcode_sample, code_set = "all_codes")
}
\arguments{
\item{loxcode_experiment}{loxcode_experiment object}

\item{loxcode_sample}{A matrix of the selected samples}

\item{code_set}{A matrix of the selected code sets}
}
\value{
Saturation plot of the loxcode sample
}
\description{
Saturation plot
}
