% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{setup_metadata}
\alias{setup_metadata}
\alias{setup_metadata,loxcode_experiment-method}
\title{Generate metadata for the loxcode experiment}
\usage{
setup_metadata(x)

\S4method{setup_metadata}{loxcode_experiment}(x)
}
\arguments{
\item{x}{loxcode object to be updated}
}
\description{
Generate metadata for the loxcode experiment
}
