% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{shiny_running}
\alias{shiny_running}
\title{CHECK IF IN SHINY}
\usage{
shiny_running()
}
\value{
True if shiny::runApp is round in callstack
}
\description{
CHECK IF IN SHINY
}
\examples{
# Load the package

# Check if the current environment is within a Shiny app
shiny_running()
}
