% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_nsamples}
\alias{simulate_nsamples}
\alias{simulate_nsamples,lox_casette-method}
\title{Simulate Loxcode Samples}
\usage{
simulate_nsamples(
  lox,
  nsamples = 10,
  ncodes = 3000,
  dist_type = "Poisson",
  npois = NULL,
  ref = NULL,
  name = "simulation"
)

\S4method{simulate_nsamples}{lox_casette}(
  lox,
  nsamples = 10,
  ncodes = 300,
  dist_type = "Poisson",
  npois = NULL,
  ref = NULL,
  name = "simulation"
)
}
\arguments{
\item{lox}{original casette}

\item{nsamples}{number of loxcode_samples}

\item{ncodes}{number of casettes/ number of cells in which a barcode is made}

\item{dist_type}{model from existing sample or poisson distribution}

\item{npois}{mean number of recombination events per casette}

\item{ref}{reference sample from which distribution of dist_orig is obtained}

\item{name}{sample_name}
}
\value{
a list of loxcode_sample with simulated data
}
\description{
Simulate Loxcode Samples
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assuming lox is defined and has the required structure
# simulate_nsamples(lox, nsamples = 5, ncodes = 3, dist_type = "sample", npois = 10, ref = reference_sample, name = "Sample1")
}
