% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_reads}
\alias{simulate_reads}
\title{Simulate recombination of multiple casettes}
\usage{
simulate_reads(lox, c, n)
}
\arguments{
\item{lox}{original casette}

\item{c}{number of casettes}

\item{n}{number of recombination events per casette}
}
\value{
list of casettes after cre lox
}
\description{
Simulate recombination of multiple casettes
}
