% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_runs}
\alias{simulate_runs}
\title{Simulate n lox recombination events}
\usage{
simulate_runs(lox, n)
}
\arguments{
\item{lox}{vector of lox sites}

\item{n}{number of recombination events to run}
}
\value{
vector of lox sites after n recombinations
}
\description{
Simulate n lox recombination events
}
