% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_sample}
\alias{simulate_sample}
\title{Simulate a loxcode_sample}
\usage{
simulate_sample(lox = NULL, c = 3000, n = 3, ref = NULL, name = "simulation")
}
\arguments{
\item{lox}{original casette}

\item{c}{number of casettes/ number of cells in which a barcode is made}

\item{n}{mean number of recombination events per casette}

\item{ref}{reference sample from which distribution of dist_orig is obtained}

\item{name}{sample_name}
}
\value{
a loxcode_sample with simulated data
}
\description{
Simulate a loxcode_sample
}
