% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{size_plot,loxcode_sample-method}
\alias{size_plot,loxcode_sample-method}
\title{Plots distribution of sizes found in samples}
\usage{
\S4method{size_plot}{loxcode_sample}(x)
}
\description{
Plots distribution of sizes found in samples
}
