% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{size_plot}
\alias{size_plot}
\title{Plots distribution of loxcode sizes found in sample}
\usage{
size_plot(x)
}
\arguments{
\item{x}{loxcode_sample object}
}
\description{
Plots distribution of loxcode sizes found in sample
}
