% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stats_bar_plot}
\alias{stats_bar_plot}
\title{stats bar plot
helper function for readstats_plot to generate size and complexity plots}
\usage{
stats_bar_plot(lox, counts, codes, count_matrix, plot, pos, labels)
}
\arguments{
\item{lox}{loxcode object}

\item{counts}{count_matrix selected}

\item{codes}{code_set selected}

\item{count_matrix}{name of count_matrix selected}

\item{plot}{plot type (size/complexity)}

\item{pos}{fill type}

\item{labels}{specify if samples should be labelled by sample_name or alias}
}
\value{
ggplot object of stats bar plot
}
\description{
stats bar plot
helper function for readstats_plot to generate size and complexity plots
}
