% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sets.R
\name{summary_table}
\alias{summary_table}
\title{Summary table (v3)}
\usage{
summary_table(lox, sample_set = "all_samples")
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{sample_set}{sample set}
}
\value{
table of sample information
}
\description{
Summary table (v3)
}
