% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{summary_table1}
\alias{summary_table1}
\title{Summary Table}
\usage{
summary_table1(x, c = "all_samples")
}
\arguments{
\item{x}{loxcode_experiment object}

\item{c}{count_matrix of samples to display}
}
\value{
A table of the selected samples
}
\description{
Tabulates the barcode counts, total number of reads, maximum complexity and consensus filtered for each sample in an experiment
}
