% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{summary_table2}
\alias{summary_table2}
\alias{summary_table2,loxcode_experiment-method}
\title{Summary table (v2)}
\usage{
summary_table2(lox, s = "all_samples")

\S4method{summary_table2}{loxcode_experiment}(lox, s = "all_samples")
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{s}{sample set}
}
\value{
table of sample information
}
\description{
Summary table (v2)
}
\examples{
# Load required packages
library(LoxCodeR2024)

# Example usage
# Assuming lox and s are defined
summary_table2(lox, s)
}
