% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\docType{methods}
\name{validate,loxcode_sample-method}
\alias{validate,loxcode_sample-method}
\title{Add cassette validation column to decoded cassette data}
\usage{
\S4method{validate}{loxcode_sample}(x)
}
\description{
Add cassette validation column to decoded cassette data
}
