% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{venn_3way}
\alias{venn_3way}
\title{3-way Venn diagram}
\usage{
venn_3way(x, a, b, c, size_range, dist_range, ...)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{a}{sample A name}

\item{b}{sample B name}

\item{c}{sample C name}

\item{size_range}{range of loxcode sizes to show in Venn diagram. Both lower and upper bounds are inclusive}

\item{dist_range}{range of dist_orig values to show in Venn diagram. Both lower and upper bounds are inclusive}

\item{labels}{optional list of custom labels for samples A, B, C, otherwise sample names for A, B, C are used.}
}
\value{
grobTree containing the 3-way Venn diagram graphic
}
\description{
Plot 3-way Venn diagram showing common and distinct barcodes between two given samples in an experiment using VennDiagram package
}
