% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barcode_table}
\alias{barcode_table}
\alias{barcode_table,loxcode_experiment-method}
\title{Generate bar code table}
\usage{
barcode_table(
  lox,
  count_matrix = "all_samples",
  code_set = "all_codes",
  labels = "alias"
)

\S4method{barcode_table}{loxcode_experiment}(
  lox,
  count_matrix = "all_samples",
  code_set = "all_codes",
  labels = "sample"
)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{count_matrix}{A matrix of the selected samples}

\item{code_set}{code set of the selected sample}

\item{labels}{specify if graph should be labelled by sample name or alias}
}
\value{
A table containing the data about the bar codes
}
\description{
A function used to generate table for the barcode details.
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assume lox, code_set, and count_matrix are defined
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
barcode_table(lox, code_set= "all_codes", count_matrix= "all_samples", labels = "alias")
}
