% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{code_freq_table}
\alias{code_freq_table}
\alias{code_freq_table,loxcode_experiment-method}
\title{Code frequency table}
\usage{
code_freq_table(x, s = "all_samples", c = "all_codes")

\S4method{code_freq_table}{loxcode_experiment}(x, s = "all_samples", c = "all_codes")
}
\arguments{
\item{x}{loxcode experiment object}

\item{s}{A matrix of the selected samples}

\item{c}{A matrix of the selected code sets}
}
\value{
data frame of code frequencies
}
\description{
Creates a data.frame of the frequency for each size and complexity combination
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assume x, s, and c are defined
x <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")
code_freq_table(x, s="all_samples", c="all_codes")

}
