% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{code_frequency_pie}
\alias{code_frequency_pie}
\alias{code_frequency_pie,loxcode_experiment-method}
\title{Code frequency pie}
\usage{
code_frequency_pie(x, s = "all_samples", c = "all_codes")

\S4method{code_frequency_pie}{loxcode_experiment}(x, s = "all_samples", c = "all_codes")
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{sample set of independent samples}

\item{c}{A matrix of the selected code sets}
}
\value{
scatterpie plot of the codes in every permutation of size and complexity
}
\description{
Display the frequency proportions of codes of every permutation of size and complexity
}
\examples{
# Load required packages
library(loxcodeR)
library(scatterpie)

# Example usage
# Assuming x, s, and c are defined and have the required structure
x <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/Week2.rds")
code_frequency_pie(x, s="all_samples", c="all_codes")
}
