% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{codeset_table}
\alias{codeset_table}
\alias{codeset_table,loxcode_experiment-method}
\title{codeset table}
\usage{
codeset_table(x, n)

\S4method{codeset_table}{loxcode_experiment}(x, n)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{n}{name of codeset to display}
}
\value{
dataframe of codeset information
}
\description{
Tabulates the codes in a codeset
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assume x and n are defined
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
x <- readRDS(datards)
n="all_codes"
codeset_table(x, n)

}
