% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sets.R
\name{collapse}
\alias{collapse}
\alias{collapse,loxcode_experiment-method}
\title{Collapses the count matrix into samples with the sample metadata (v2)}
\usage{
collapse(lox, count_matrix, collapse, name, union = TRUE, average = FALSE)

\S4method{collapse}{loxcode_experiment}(lox, count_matrix, collapse, name, union = TRUE, average = FALSE)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{count_matrix}{current count_matrix}

\item{collapse}{column names of metadata on which to collapse}

\item{name}{name of new count_matrix}

\item{union}{boolean, True if barcodes in either should be counted}

\item{average}{boolean, True if counts should be averaged instead of summed}
}
\value{
loxcode_experiment object with new collapsed samples
}
\description{
Collapses the count matrix into samples with the sample metadata (v2)
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assuming lox, count_matrix, collapse, name, union, and average are defined and have the required structure
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
count_matrix="all_samples"
name="new"
collapse="Organ"
collapse(lox, count_matrix, collapse, name = "new_count_matrix", union = TRUE, average = FALSE)
}
