% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plot.R
\name{correlation_plot}
\alias{correlation_plot}
\title{correlation plot}
\usage{
correlation_plot(
  lox,
  count_matrix = "all_samples",
  code_set = "all_codes",
  method_ = "pearson"
)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{count_matrix}{A matrix of the selected samples}

\item{code_set}{A matrix of the selected code sets}

\item{method_}{correlation method (p)}

\item{parameter}{metadata type to display}

\item{data}{metadata to display}
}
\description{
Function used to create a correlation plot
}
