% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decode}
\alias{decode}
\title{Decode FASTQ
Recover loxcodes from raw Illumina FASTQ output}
\usage{
decode(r, name, meta, min_r1_len, min_r2_len, full, sat)
}
\arguments{
\item{r}{Paths of R1, 2 respectively}

\item{name}{Name of the Fastq file}

\item{meta}{User-defined data-frame for sample metadata}

\item{min_r1_len}{Minimum length of r1}

\item{min_r2_len}{Minimum length of r2}

\item{full}{whether to supply full output (including read IDs, etc)}

\item{sat}{Satiration amount}
}
\value{
S4 loxcode_sample object with decoded results
}
\description{
Decode FASTQ
Recover loxcodes from raw Illumina FASTQ output
}
\examples{
# Example usage:
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
}
