% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decode}
\alias{decode}
\title{Decode FASTQ}
\usage{
decode(r, name, meta, min_r1_len, min_r2_len)
}
\arguments{
\item{r}{Paths of R1, 2 respectively}

\item{meta}{User-defined data-frame for sample metadata}

\item{min_read_length}{min read length for R1, R2 filter respectively}
}
\value{
S4 loxcode_sample object with decoded results
}
\description{
Recover loxcodes from raw Illumina FASTQ output
}
