% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decode}
\alias{decode}
\title{Decode FASTQ}
\usage{
decode(prefix, dir, meta)
}
\arguments{
\item{prefix}{Prefix of sample name}

\item{dir}{Directory where FASTQ files are stored.}

\item{meta}{User-defined data-frame for sample metadata}
}
\description{
Recover loxcodes from raw Illumina FASTQ output
decode() will search for samples with names of the form <prefix>_R{1, 2}_001.fastq
}
