% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\docType{class}
\name{decode_output-class}
\alias{decode_output-class}
\alias{decode_output}
\title{S4 class to contain output of decode()}
\description{
S4 class to contain output of decode()
}
\section{Slots}{

\describe{
\item{\code{data}}{contains raw output of decode() as a data.frame}

\item{\code{read_ids}}{list of FASTQ}
}}

