% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{delete_count_matrix}
\alias{delete_count_matrix}
\alias{delete_count_matrix,loxcode_experiment-method}
\title{Creates a new count_matrix with dataframe of samples}
\usage{
delete_count_matrix(x, c)

\S4method{delete_count_matrix}{loxcode_experiment}(x, c)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{c}{name of existing count_matrix to be deleted}
}
\value{
updated loxcode_experiment object
}
\description{
Creates a new count_matrix with dataframe of samples
}
\examples{
# Load a sample loxcode_experiment object
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")


# Delete the existing count_matrix named 'count_matrix_1'
lox_updated <- delete_count_matrix(lox, 'all_sample')
}
