% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dist_orig_plot}
\alias{dist_orig_plot}
\alias{dist_orig_plot,loxcode_experiment-method}
\title{Dist_orig Plot}
\usage{
dist_orig_plot(
  lox,
  sample,
  count_matrix = "all_samples",
  code_set = "all_codes",
  labels = "alias"
)

\S4method{dist_orig_plot}{loxcode_experiment}(
  lox,
  sample,
  count_matrix = "all_samples",
  code_set = "all_codes",
  labels = "alias"
)
}
\arguments{
\item{lox}{loxcode experiment object of current experiment}

\item{sample}{loxcode sample object}

\item{count_matrix}{A matrix of the selected samples}

\item{code_set}{A matrix of the selected code sets}

\item{labels}{A string containing sample name or alias}
}
\value{
a plot of the distribution of distance from origin in the sample
}
\description{
Plots distribution of distance from origin found in sample
}
\examples{
# Load a sample loxcode experiment object and sample
lox <- readRDS("~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds")

# Plot the distribution of distance from origin
dist_orig_plot(lox, sample="Sample 1", count_matrix= "all_samples", code_set= "all_codes", labels = "alias")
}
