% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{excise}
\alias{excise}
\title{Perform an excision on the sequence}
\usage{
excise(lox, a, b)
}
\arguments{
\item{lox}{lox_casette object}

\item{a}{lox_site position 1}

\item{b}{lox_site position 2}
}
\value{
an excised lox_casette object
}
\description{
Perform an excision on the sequence
}
