% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{exp_table}
\alias{exp_table}
\alias{exp_table,list-method}
\title{Experiments Table}
\usage{
exp_table(exp)

\S4method{exp_table}{list}(exp)
}
\arguments{
\item{exp}{list of loxcode experiment objects}
}
\value{
a data frame containing the information of loxcode experiment
}
\description{
Presents a list of loxcode experiments in table form
}
\examples{
# Load required packages
library(loxcoder)

# Example usage
# Assume exp is a list of loxcode experiment objects
lox <- readRDS("~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds")
exp_table(c(lox))

}
