% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{exp_table}
\alias{exp_table}
\title{Experiments Table}
\usage{
exp_table(exp)
}
\arguments{
\item{exp}{list of loxcode experiment objects}
}
\value{
a data frame containing the information of loxcode experiment
}
\description{
Presents a list of loxcode experiments in table form
}
