% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{fill_alias}
\alias{fill_alias}
\alias{fill_alias,loxcode_experiment-method}
\title{Generate alias for the loxcode experiment samples}
\usage{
fill_alias(x, ...)

\S4method{fill_alias}{loxcode_experiment}(x)
}
\arguments{
\item{x}{loxcode experiment sample}

\item{...}{value of alias to fill}
}
\value{
loxcode object with alias
}
\description{
Generate alias for the loxcode experiment samples
}
\examples{
# Load necessary libraries and data
library(loxcodeR)
lox <- readRDS('~/Desktop/loxcodeR/LoxcodeR_app/Week2.rds')


# Generate alias for the samples
lox_experiment <- fill_alias(lox)
}
