% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gatherpairs}
\alias{gatherpairs}
\title{Gather Pairs}
\usage{
gatherpairs(
  data,
  ...,
  xkey = ".xkey",
  xvalue = ".xvalue",
  ykey = ".ykey",
  yvalue = ".yvalue",
  na.rm = FALSE,
  convert = FALSE,
  factor_key = FALSE
)
}
\description{
Function used to
}
