% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{generate_alias}
\alias{generate_alias}
\alias{generate_alias,loxcode_experiment-method}
\title{generate sample aliases}
\usage{
generate_alias(lox, s = "all_samples", meta)

\S4method{generate_alias}{loxcode_experiment}(lox, s = "all_samples", meta)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{s}{sample set}

\item{meta}{columns of meta data on which to collapse}
}
\value{
loxcode_experiment object with aliases generated
}
\description{
generate sample aliases
}
\examples{
# Load necessary libraries and data
library(loxcodeR)

# Create a loxcode experiment object
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")

# Generate sample aliases
lox_experiment <- generate_alias(lox, s = 'all_samples', 'tissue')
}
