% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_alias}
\alias{get_alias}
\title{Get sample alias}
\usage{
get_alias(lox, set, name)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{set}{sample set name}

\item{name}{sample_name}
}
\value{
sample alias
}
\description{
Get sample alias
}
