% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_alias}
\alias{get_alias}
\alias{get_alias,loxcode_experiment-method}
\title{Get sample alias}
\usage{
get_alias(lox, set, name)

\S4method{get_alias}{loxcode_experiment}(lox, set, name)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{set}{sample set name}

\item{name}{sample_name}
}
\value{
sample alias
}
\description{
Get sample alias
}
\examples{
# Load necessary libraries and data
library(loxcodeR)
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")

# Get sample alias
alias <- get_alias(lox,"N712_N501","all_sample")
}
