% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_barcode_stats_rep}
\alias{get_barcode_stats_rep}
\title{Barcode Stats}
\usage{
get_barcode_stats_rep(union_bc, rep)
}
\arguments{
\item{union_bc}{Barcodes data to check in a sample}

\item{rep}{loxcode object containing data of a sample}
}
\value{
A object containing the barcode and data on the matched barcode in the sample
}
\description{
Number of times a barcode is seen in a sample
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assuming union_bc and rep are defined
rep1 <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")@samples$N712_N501
rep2 <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")@samples$N712_N506
bc_union <- barcode_union(rep1, rep2, type="size", range=c(1, 13))
get_barcode_stats_rep(bc_union, rep1)
}
