% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_cass_vec}
\alias{get_cass_vec}
\title{Convert vector of cassette strings to list of numeric vectors}
\usage{
get_cass_vec(c)
}
\arguments{
\item{c}{vector of cassette strings}
}
\description{
Convert vector of cassette strings to list of numeric vectors
}
