% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sets.R
\name{get_collapsed_meta}
\alias{get_collapsed_meta}
\title{Get metadata of a merged experiment}
\usage{
get_collapsed_meta(x, s)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{set of merged samples}
}
\value{
a data frame of meta data
}
\description{
Get metadata of a merged experiment
}
