% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sets.R
\name{get_collapsed_meta}
\alias{get_collapsed_meta}
\alias{get_collapsed_meta,loxcode_experiment-method}
\title{Get metadata of a merged experiment}
\usage{
get_collapsed_meta(x, s)

\S4method{get_collapsed_meta}{loxcode_experiment}(x, s)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{set of merged samples}
}
\value{
a data frame of meta data
}
\description{
Get metadata of a merged experiment
}
\examples{
# Load necessary libraries and data
library(loxcodeR)

# Create a loxcode experiment object
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)

# Get collapsed metadata
meta <- get_collapsed_meta(lox,"all_sample")
}
