% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{get_collapsed_meta2}
\alias{get_collapsed_meta2}
\alias{get_collapsed_meta2,loxcode_experiment-method}
\title{Get metadata of a merged experiment (old version)}
\usage{
get_collapsed_meta2(x, s)

\S4method{get_collapsed_meta2}{loxcode_experiment}(x, s)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{set of merged samples}
}
\value{
a data frame of meta data
}
\description{
Get metadata of a merged experiment (old version)
}
