% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{get_data}
\alias{get_data}
\alias{get_data,loxcode_experiment-method}
\title{Get unvalidated readout data for a given sample}
\usage{
get_data(x, s)

\S4method{get_data}{loxcode_experiment}(x, s)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{sample name}
}
\value{
data.frame containing unvalidated readout data for sample s
}
\description{
Equivalent to performing LoxCodeR2024::data() on the sample s
}
\examples{
# Load necessary libraries and data
library(LoxCodeR2024)
data <- read.csv("your_data.csv")

# Create a loxcode experiment object
lox_experiment <- create_lox_experiment(data)

# Get unvalidated readout data for a sample
sample_data <- get_data(lox_experiment, "Sample_Name")
}
