% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\name{get_origin_dist}
\alias{get_origin_dist}
\alias{get_origin_dist,loxcode_sample-method}
\title{Fetch distances from origin (dist_orig)}
\usage{
get_origin_dist(x)

\S4method{get_origin_dist}{loxcode_sample}(x)
}
\arguments{
\item{x}{loxcode object}
}
\value{
Distance from origin of valid cassette
}
\description{
Appends a column of dist_orig values for valid cassettes
}
\examples{
# Load necessary libraries and data
library(loxcoder)
lox <- readRDS("~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds")

# Fetch distances from origin
#dist_orig_values <- get_origin_dist(lox)
}
