% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{get_pair_dist}
\alias{get_pair_dist}
\title{Get pairwise distance}
\usage{
get_pair_dist(codesA, codesB)
}
\arguments{
\item{codesA}{data.frame}

\item{codesB}{data.frame}
}
\value{
Pairwise distance between code A and B
}
\description{
Funtion to get pairwise distance of two codesets
}
\examples{
# Load necessary libraries and data
library(loxcoder)
lox <- readRDS("~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds")

# Assuming you have two codesets
#codeset_A <- data[data$codeset == "A", ]
#codeset_B <- data[data$codeset == "B", ]

# Get pairwise distance
#pairwise_distance <- get_pairwise_distance(codeset_A, codeset_B)
}
