% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stats_plot.R
\name{get_ratio_plot}
\alias{get_ratio_plot}
\title{get ratio plot
helper function to generate plot of ratios}
\usage{
get_ratio_plot(lox, counts, codes, count_matrix, labels)
}
\arguments{
\item{lox}{loxcode_experiment}

\item{counts}{count_matrix selected}

\item{codes}{code_set selected}

\item{count_matrix}{name of count_matrix selected}

\item{samples}{list of samples by name}
}
\description{
get ratio plot
helper function to generate plot of ratios
}
