% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\name{get_rec_prob}
\alias{get_rec_prob}
\title{Get recombination distance distribution}
\usage{
get_rec_prob(x, size)
}
\description{
Returns the proportion of valid cassettes for each valid recombination distance (dist_orig)
}
