% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_samplename}
\alias{get_samplename}
\title{Get sample name from alias}
\usage{
get_samplename(lox, set, name)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{set}{sample set name}

\item{name}{alias}
}
\value{
sample alias
}
\description{
Get sample name from alias
}
