% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{get_valid}
\alias{get_valid}
\alias{get_valid,loxcode_experiment-method}
\title{Get validated readout data for a given sample}
\usage{
get_valid(x, s)

\S4method{get_valid}{loxcode_experiment}(x, s)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{sample name}
}
\value{
loxcode object with valid data
}
\description{
Equivalent to performing loxcodeR::valid() on the sample s
}
\examples{
# Load necessary libraries and data
library(loxcodeR)
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)

valid_data <- get_valid(lox, 'N712_N501')
}
