% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{get_valid}
\alias{get_valid}
\alias{get_valid,loxcode_experiment-method}
\title{Get validated readout data for a given sample}
\usage{
get_valid(x, s)

\S4method{get_valid}{loxcode_experiment}(x, s)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{sample name}
}
\value{
loxcode object with valid data
}
\description{
Equivalent to performing LoxCodeR2024::valid() on the sample s
}
\examples{
# Load necessary libraries and data
library(LoxCodeR2024)
data <- read.csv("your_data.csv")

# Assuming you have a loxcode_experiment object called lox
sample_name <- "your_sample_name"  # Specify the name of the sample
valid_data <- get_valid(lox, sample_name)
}
