% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getData}
\alias{getData}
\title{Get sample data}
\usage{
getData(sample, plot_type = "size")
}
\arguments{
\item{sample}{loxcode sample object}

\item{plot_type}{"size" or "dist_orig"}
}
\value{
data frame with sizes in a sample
}
\description{
Get a list of sizes from the sample
}
\examples{
# Load required packages
library(LoxCodeR2024)

# Example usage
# Assuming sample is defined
get_sample_data(sample, plot_type = "size")
}
