% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impute_13}
\alias{impute_13}
\title{Impute missing elements}
\usage{
impute_13(c, sizes)
}
\arguments{
\item{c}{list of loxcode numbers present in each cassete}

\item{sizes}{sizes of loxcode}
}
\value{
13-element cassette
}
\description{
Impute missing elements
}
\examples{
# Example usage:
# Assuming 'cassette_list' is a list of loxcode numbers present in each cassette,
# and 'loxcode_sizes' is a vector containing the sizes of the loxcode.
# Impute the missing elements in each cassette to form a 13-element cassette
lox <- readRDS("~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds")
#cassette <- impute_13(cassette_list, loxcode_sizes)
#cassette
}
