% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{load_from_xlsx}
\alias{load_from_xlsx}
\title{Create experiment}
\usage{
load_from_xlsx(name, s, dir, load = TRUE, full = FALSE, sat = FALSE)
}
\arguments{
\item{name}{string, name of the experiment}

\item{s}{string, path to user-specified Excel file containing experiment samples.}

\item{dir}{string, path to directory containing fastq files}

\item{load}{boolean, whether to load samples or not (default is TRUE)}

\item{full}{boolean, whether to produce full debugging output (default is FALSE, this uses significantly more memory)}

\item{sat}{boolean, whether to keep saturation info}
}
\value{
loxcode_experiment object
}
\description{
Create a loxcode_experiment from xlsx sample table.
}
\details{
Must contain the following columns:

* `sample` -- sample name. This is used as the unique identifier for the sample

* `prefix` -- sample prefix. suffix_R1 and suffix_R2 are appended to this

* `meta` -- sample metadata. This is user specified

* `min_r1_len` -- minimum R1 length

* `min_r2_len` -- minimum R2 length
}
\examples{
# Example usage:
name <- "MyExperiment"
s <- "path/to/excel_file.xlsx"
dir <- "path/to/fastq_directory"
#experiment <- create_experiment(name, s, dir)
#experiment
}
