% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\name{load_pair_distmaps}
\alias{load_pair_distmaps}
\title{Load distance maps for pairwise distances}
\usage{
load_pair_distmaps(path)
}
\arguments{
\item{path}{path to directory containing the maps 0 (for 13-element distances) and 1 (for 9-element distances)}
}
\value{
distance maps files for pairwise distance
}
\description{
Load distance maps for pairwise distances
}
\examples{
# Example usage:
#path <- "path/to/your/directory"
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
#pairwise_distance_maps <- load_distance_maps(path)
#pairwise_distance_maps
}
