% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\name{load_prob_files}
\alias{load_prob_files}
\title{Load probability tables (Markov chain)}
\usage{
load_prob_files(path)
}
\arguments{
\item{path}{path to directory containing the Markov chain tables}
}
\value{
Probability table
}
\description{
Load probability tables (Markov chain)
}
\examples{
# Example usage:
#path <- "path/to/markov_chain_tables_directory"
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")
#probability_table <- load_probability_tables(path)
#probability_table
}
