% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{load_samples}
\alias{load_samples}
\alias{load_samples,loxcode_experiment-method}
\title{Load data from FASTQ for all samples in a loxcode_experiment object}
\usage{
load_samples(x, ...)

\S4method{load_samples}{loxcode_experiment}(x, full = FALSE, sat = FALSE)
}
\arguments{
\item{x}{loxcode_experiment object for which to load samples}

\item{...}{Used to pass variable name for loading data}

\item{full}{description}

\item{sat}{desc}
}
\value{
loxcode_experiment object with sample data loaded
}
\description{
For each sample in the loxcode_experiment, barcodes readout is performed from FASTQ
followed by element imputation for the 13-element codes. Loxcodes are then validated, IDs fetched,
and distance-from-origin is retrieved.
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object
lox <- readRDS('~/Desktop/loxcodeR/LoxcodeR_app/Week2.rds')
experiment <- load_samples(lox)
experiment
}
