% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\docType{class}
\name{loxcode_experiment-class}
\alias{loxcode_experiment-class}
\title{S4 class for a single loxcode experiment, enabling handling of multiple samples}
\description{
S4 class for a single loxcode experiment, enabling handling of multiple samples
}
\section{Slots}{

\describe{
\item{\code{name}}{string, name of the experiment}

\item{\code{suffix_R1}}{string, R1 suffix}

\item{\code{suffix_R2}}{string, R2 suffix}

\item{\code{dir}}{string, directory containing R1, R2 fastq files}

\item{\code{samples}}{list, contains loxcode_sample objects}

\item{\code{samp_table}}{data.frame, user-specified table.}
}}

