% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\docType{class}
\name{loxcode_sample-class}
\alias{loxcode_sample-class}
\alias{loxcode_sample}
\title{S4 class to represent loxcode experimental sample data}
\value{
Loxcode experiment sample data
}
\description{
S4 class to represent loxcode experimental sample data
}
\section{Slots}{

\describe{
\item{\code{decode}}{A data.frame to contain raw decode data from loxcodeR::decode()}

\item{\code{meta}}{A data.frame for user-defined sample metadata}

\item{\code{name}}{Contains name of the experiment}

\item{\code{files}}{A vector containing file names of loxcodeR experiment}

\item{\code{decode_stats}}{A list containing statistics of decoded data}

\item{\code{consensus_filtered_data}}{A vector containing filtered data}
}}

\examples{
# Define sample metadata
sample_meta <- data.frame(
  sample_name = "Sample1",
  meta_value1 = 123,
  meta_value2 = "abc"
)

# Create a LoxcodeExperimentSample object
#sample_data <- loxcode_sample(
 # decode = data.frame(...),  # Raw decode data
 # meta = sample_meta,        # Sample metadata
 # name = "Sample1",          # Name of the experiment
 # files = c("file1.fastq", "file2.fastq"),  # File names
 # decode_stats = list(...),  # Decode statistics
 # consensus_filtered_data = c(...)  # Filtered data
#)
}
