% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{make_codeset_index}
\alias{make_codeset_index}
\title{Creates a new codeset with dataframe of codesets testing}
\usage{
make_codeset_index(x, c, I, n)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{c}{name of existing codeset to choose from}

\item{I}{indices to include}

\item{n}{name of the new codeset}
}
\value{
new loxcode_experiment object with a new codeset
}
\description{
Creates a new codeset with dataframe of codesets testing
}
