% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{make_filtered_codeset}
\alias{make_filtered_codeset}
\alias{make_filtered_codeset,loxcode_experiment-method}
\title{Create new filtered code set according to defined parameters}
\usage{
make_filtered_codeset(x, s, c, t, m, n)

\S4method{make_filtered_codeset}{loxcode_experiment}(x, s, c, t, m, n)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{independent sample set}

\item{c}{code set}

\item{t}{tolerance threshold for repeated proportions}

\item{m}{maximum repeats tolerated}

\item{n}{new code set name}
}
\value{
new loxcode_experiment object
}
\description{
Create new filtered code set according to defined parameters
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object
# Create a new filtered code set named 'filtered_codeset' based on parameters
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")
new_experiment <- make_filtered_codeset(lox,'all_samples','all_codes',0.1,3,'test')
new_experiment
}
