% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\name{makeid}
\alias{makeid}
\alias{makeid,loxcode_sample-method}
\title{Get cassette IDs}
\usage{
makeid(x)

\S4method{makeid}{loxcode_sample}(x)
}
\arguments{
\item{x}{loxcode data object}
}
\value{
cassette Id
}
\description{
Appends a column of packed cassette IDs, or -1 if it cannot be packed
}
\examples{
# Example usage:
# Assuming 'lox_data' is a loxcode data object
# Get cassette IDs and append them as a new column
lox <- readRDS("~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds")
cassette_ids <- makeid(lox@samples$N709_N501)
cassette_ids
}
