% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{merge_by}
\alias{merge_by}
\alias{merge_by,loxcode_experiment-method}
\title{Merge samples by label}
\usage{
merge_by(x, by)

\S4method{merge_by}{loxcode_experiment}(x, by)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{by}{name of column in samp_table to merge by}
}
\value{
loxcode_experiment object containing merged samples
}
\description{
Merge samples present in a loxcode_experiment object according to a specified label column in
samp_table. All samples with a common value for the label column are merged together. The resulting
loxcode_experiment object contains loxcode_samples named by the corresponding labels in by.
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object and 'label_column' is the name of the column to merge by
# Merge samples based on the values in the 'label_column'
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
# merged_experiment <- merge_by(lox, by = 'sample')
# merged_experiment
}
