% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{merge_experiments}
\alias{merge_experiments}
\alias{merge_experiments,loxcode_experiment-method}
\title{Merge two experiments}
\usage{
merge_experiments(lox1, lox2, name = NULL)

\S4method{merge_experiments}{loxcode_experiment}(lox1, lox2, name = NULL)
}
\arguments{
\item{lox1}{loxcode_experiment 1}

\item{lox2}{loxcode_experiment 2}

\item{name}{experiment name}
}
\value{
new merged loxcode_experiment
}
\description{
Merge two experiments
}
\examples{
# Example usage:
# Assuming 'experiment1' and 'experiment2' are loxcode_experiment objects
# Merge the two experiments into a new experiment named 'merged_experiment'
lox <- readRDS('~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds')
merged_experiment <- merge_experiments(lox,lox,'test')
merged_experiment
}
