% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{merge_experiments2}
\alias{merge_experiments2}
\title{Merge a vector of experiments}
\usage{
merge_experiments2(experiments, name)
}
\arguments{
\item{experiments}{vector of loxcode_experiment objects}

\item{name}{name of new loxcode_experiment object}
}
\value{
loxcode_experiment object
}
\description{
Merge a vector of experiments
}
\examples{
# Example usage:
# Assuming 'experiment1', 'experiment2', and 'experiment3' are loxcode_experiment objects
# Merge the three experiments into a new experiment named 'merged_experiment'
lox <- readRDS('~/Desktop/loxcodeR/LoxcodeR_app/Week2.rds')
merged_experiment <- merge_experiments2(c(lox,lox),'test')
merged_experiment
}
