% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{merge_sample}
\alias{merge_sample}
\title{Merge two loxcode_sample objects}
\usage{
merge_sample(s1, s2)
}
\arguments{
\item{s1}{sample 1}

\item{s2}{sample 2}
}
\value{
loxcode_sample object corresponding to merged sample
}
\description{
Barcodes from s1 and s2 are merged. For barcodes present in both samples, read counts are
summed. Sample descriptors and other members are concatenated in order.
}
