% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sets.R
\name{merge_sample_list}
\alias{merge_sample_list}
\alias{merge_sample_list,loxcode_experiment-method}
\title{Merge list of samples
Merges a list of samples into one loxcode_sample object}
\usage{
merge_sample_list(lox, samples, union = TRUE, average = FALSE)

\S4method{merge_sample_list}{loxcode_experiment}(lox, samples, union = TRUE, average = FALSE)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{samples}{list of sample names to be merged}

\item{union}{True if barcodes in either samples should be counted}

\item{average}{True if counts should be averaged instead of summed}
}
\value{
merged loxcode_sample object
}
\description{
Merge list of samples
Merges a list of samples into one loxcode_sample object
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object and 'sample_list' is a list of sample names to be merged
# Merge the samples in 'sample_list' into a single loxcode_sample object, summing their counts
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
# merged_sample <- merge_sample_list(lox = experiment, samples = sample_list, union = TRUE, average = FALSE)
# merged_sample
}
