% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{min_flip_dist}
\alias{min_flip_dist}
\title{Distance at which the flip happened}
\usage{
min_flip_dist(c, size, v)
}
\arguments{
\item{c}{list of loxcode numbers present in each cassette}

\item{size}{sizes of loxcode}

\item{v}{flip distance}
}
\description{
Distance at which the flip happened
}
