% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R, R/loxcode_sample.R
\name{name}
\alias{name}
\alias{name,loxcode_sample-method}
\title{Fetch experiment name}
\usage{
name(x)

name(x)

\S4method{name}{loxcode_sample}(x)
}
\arguments{
\item{x}{loxcode_experiment object}
}
\value{
name of experiment
}
\description{
Fetch experiment name

Get loxcode sample name (description)
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object
# Retrieve the name of the experiment
lox <- readRDS("~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds")
#experiment_name <- name(lox)
#experiment_name
# Example usage:
# Assuming 'experiment' is a loxcode_sample object
# Retrieve the name of the experiment
# experiment_name <- name(experiment)
# experiment_name
}
