% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{new_alias}
\alias{new_alias}
\alias{new_alias,loxcode_experiment-method}
\title{Changes the alias of a sample}
\usage{
new_alias(x, c, s, n)

\S4method{new_alias}{loxcode_experiment}(x, c, s, n)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{c}{current count_matrix}

\item{s}{sample name}

\item{n}{new alias}
}
\value{
loxcode_experiment object with changed sample alias
}
\description{
Changes the alias of a sample
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object
# Change the alias of the sample named 'old_sample' to 'new_sample'
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
new_experiment <- new_alias(lox,'all_samples','N712_N501','test')
new_experiment
}
