% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pack}
\alias{pack}
\title{Pack cassettes into cassette ID}
\usage{
pack(c, v)
}
\arguments{
\item{c}{a list of numeric vectors, or a character vector of decoded loxcodes}

\item{v}{vector of bool, output from validate()}
}
\value{
cassettes with cassette id
}
\description{
Pack cassettes into cassette ID
}
\examples{
# Example usage:
# Assuming 'cassette' is a list of numeric vectors representing the cassette,
# 'validity_vector' is a vector of boolean values indicating the validity of each cassette.
# Pack the cassettes into cassette IDs
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")
# cassettes_with_ID <- pack(cassette, validity_vector)
# cassettes_with_ID
}
