% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pair_comparison_plot2}
\alias{pair_comparison_plot2}
\title{Pair Comparison Plot 2}
\usage{
pair_comparison_plot2(
  lox,
  sampleset = "all_samples",
  codeset = "all_codes",
  s1,
  s2,
  colorBy = "size",
  labels = "alias",
  sizeRange = NULL,
  dist_origRange = NULL,
  firstreadRange = NULL
)
}
\arguments{
\item{lox}{Loxcode experiment object}

\item{sampleset}{A matrix of the selected samples}

\item{codeset}{A matrix of the selected code sets}

\item{s1}{First Loxcode sample}

\item{s2}{Second Loxcode sample}

\item{colorBy}{Color points by size or dist_orig or firstread}

\item{labels}{Sample name or alias as labels}
}
\value{
A plot containing pairwise comparison of the samples
}
\description{
Function to create a pair comparision plot by passing code set and sample set. This function can use custom sample and code set for plot creation.
}
