% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pair_comparison_plot_all}
\alias{pair_comparison_plot_all}
\alias{pair_comparison_plot_all,loxcode_experiment-method}
\title{Pair Comparison Plot all}
\usage{
pair_comparison_plot_all(
  lox,
  min_reads = 10,
  matrix = "all_samples",
  codeset = "all_codes"
)

\S4method{pair_comparison_plot_all}{loxcode_experiment}(
  lox,
  min_reads = 10,
  matrix = "all_samples",
  codeset = "all_codes"
)
}
\arguments{
\item{lox}{Loxcode_experiment object}

\item{min_reads}{Code set}

\item{matrix}{First Loxcode_sample}

\item{codeset}{Second Loxcode_sample}
}
\value{
A pair comparision plot comparing all samples agaist each other
}
\description{
Pair Comparison Plot all
}
\examples{
# Load required packages
library(loxcoder)

# Example usage
# Assuming lox, min_reads, matrix, and codeset are defined
lox <- readRDS("~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds")
#pair_comparison_plot_all(lox,matrix = lox@count_matrixes$`all_samples_ST114_Bcells ST114_2 NN167_Week2`,codeset = lox@code_sets$invalid_codes_ST111)
}
