% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rank_count_plot}
\alias{rank_count_plot}
\alias{rank_count_plot,loxcode_sample-method}
\title{Rank Count plot}
\usage{
rank_count_plot(x, size, ymax, dmax)

\S4method{rank_count_plot}{loxcode_sample}(x, size, ymax, dmax)
}
\arguments{
\item{x}{loxcode sample object}

\item{size}{loxcode size to consider}

\item{ymax}{max y-value for counts}

\item{dmax}{max d-value for dist_orig}
}
\value{
ggplot object of rank count plot
}
\description{
In the resulting plot, barcodes of specified size are ranked by read count. log10(count) is then plotted
against barcode rank in black. In addition, dist_orig for each barcode is shown in red with the same horizontal axis.
}
\examples{
# Load required packages
library(LoxCodeR2024)

# Example usage
# Assuming x and size are defined
rank_count_plot(x, size, ymax = 100, dmax = 50)
}
