% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{read_plot}
\alias{read_plot}
\alias{read_plot,loxcode_experiment-method}
\title{Read Plot}
\usage{
read_plot(
  lox,
  count_matrix = "all_samples",
  code_set = "all_codes",
  labels = "alias"
)

\S4method{read_plot}{loxcode_experiment}(
  lox,
  count_matrix = "all_samples",
  code_set = "all_codes",
  labels = "sample"
)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{count_matrix}{A matrix of the selected samples}

\item{code_set}{code set in selected samples}

\item{labels}{specify if graph should be labelled by sample name or alias}
}
\value{
A plot of the number of reads
}
\description{
plots the number of reads
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assuming lox, code_set, and count_matrix are defined
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
read_plot(lox)
}
