% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{readstats_plot}
\alias{readstats_plot}
\alias{readstats_plot,loxcode_experiment-method}
\title{readstats plot (v3)}
\usage{
readstats_plot(
  lox,
  count_matrix = "all_samples",
  code_set = "all_codes",
  plot = "size",
  fill = TRUE,
  labels = "alias"
)

\S4method{readstats_plot}{loxcode_experiment}(
  lox,
  count_matrix = "all_samples",
  code_set = "all_codes",
  plot = "size",
  fill = TRUE,
  labels = "sample"
)
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{count_matrix}{Count matrix of loxcode}

\item{code_set}{Code set of loxcode}

\item{plot}{specify which plot (size/complexity/both/ratio)}

\item{fill}{specify whether reads should be normalized in boxplots}

\item{labels}{specify if graph should be labelled by sample name or alias}
}
\value{
Plot based on the type of plot selected
}
\description{
Function to create size, complexity, both or ratio plot
}
\examples{
# Load required packages
library(loxcoder)

# Example usage
# Assuming lox, count_matrix, code_set, plot, and fill are defined
lox <- readRDS("~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds")
readstats_plot(lox)
}
