% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{rename_codeset}
\alias{rename_codeset}
\title{rename code set}
\usage{
rename_codeset(x, c, n)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{c}{code set}

\item{n}{new code set name}
}
\value{
new loxcode_experiment object
}
\description{
rename code set
}
