% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sets.R
\name{rename_sampleset}
\alias{rename_sampleset}
\title{rename sample set}
\usage{
rename_sampleset(x, s, n)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{sample set}

\item{n}{new code set name}
}
\value{
new loxcode_experiment object
}
\description{
rename sample set
}
