% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{retrieve_dist_origin}
\alias{retrieve_dist_origin}
\title{Assumes that input cassette is valid already!}
\usage{
retrieve_dist_origin(c, sizes)
}
\arguments{
\item{c}{input cassette}

\item{sizes}{sizes of all loxcode}
}
\value{
distance from origin
}
\description{
Assumes that input cassette is valid already!
}
\examples{
# Example usage:
# Assuming 'cassette' is a valid input cassette and 'loxcode_sizes' is a vector containing sizes of all loxcodes.
# Calculate the distance from the origin
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
#dist_from_origin <- retrieve_dist_origin(cassette, loxcode_sizes)
#dist_from_origin
}
