% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{retrieve_dist_origin_single}
\alias{retrieve_dist_origin_single}
\title{Assumes that input cassette is valid already!}
\usage{
retrieve_dist_origin_single(c, size)
}
\arguments{
\item{c}{input cassette}

\item{size}{size of single loxcode}
}
\value{
distance from origin
}
\description{
Assumes that input cassette is valid already!
}
\examples{
# Example usage:
# Assuming 'cassette' is a valid input cassette and 'loxcode_size' is the size of a single loxcode.
# Calculate the distance from the origin
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")
#dist_from_origin <- retrieve_dist_origin_single(cassette, loxcode_size)
#dist_from_origin
}
