% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{retrieve_prob}
\alias{retrieve_prob}
\title{Assumes that input cassette is valid already!}
\usage{
retrieve_prob(c, sizes, nrec)
}
\arguments{
\item{c}{input cassette}

\item{sizes}{sizes of all loxcode}

\item{nrec}{position in cassette}
}
\description{
Assumes that input cassette is valid already!
}
