% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\name{retrieve_prob_ensemble}
\alias{retrieve_prob_ensemble}
\alias{retrieve_prob_ensemble,loxcode_sample-method}
\title{Get ensemble generation probability}
\usage{
retrieve_prob_ensemble(x)

\S4method{retrieve_prob_ensemble}{loxcode_sample}(x)
}
\arguments{
\item{x}{loxcode_sample object}
}
\value{
ensemble genration probability
}
\description{
Retrieve ensemble probabilities as a weighted linear combination of Markov probabilities
where distances (dist_orig) are weighted using the sample distribution of recombination distances.
}
\details{
Results are appended to readout data as a 'prob' column.
}
\examples{
# Example usage:
# Assuming 'sample' is a loxcode_sample object
# Retrieve ensemble generation probability for the sample
lox <- readRDS("~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds")
ensemble_probability <- retrieve_prob_ensemble(lox@samples$N709_N501)
ensemble_probability
}
