% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\name{retrieve_prob_ensemble}
\alias{retrieve_prob_ensemble}
\title{Get ensemble generation probability}
\usage{
retrieve_prob_ensemble(x)
}
\arguments{
\item{x}{loxcode_sample object}
}
\description{
Retrieve ensemble probabilities as a weighted linear combination of Markov probabilities
where distances (dist_orig) are weighted using the sample distribution of recombination distances.
}
\details{
Results are appended to readout data as a 'prob' column.
}
