% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_markov}
\alias{run_markov}
\title{Calculate generation probabilities using Markov chain formulation}
\usage{
run_markov(n, To, start_idx, start_id)
}
\arguments{
\item{n}{number of steps to generate}

\item{To}{transition probabilities specified as interaction probabilities}

\item{start_idx}{size_idx from which to start}

\item{start_id}{id of cassette to initialise markov chain from (0 for 13-element)}
}
\value{
list of n+1 items, giving the distribution state at step 0, 1, ..., n. (0 for unrecombined 13-elt cassette)
}
\description{
T specifies the transition probabilities by assigning probabilities to the interactions
between LoxP sites. T must be a list of length 5 where each entry is a NumericMatrix.
T[[1]] must be a 14x14 matrix where element (i, j) = (j, i) corresponds to the probability
of interaction between LoxP sites i and j, where LoxP sites are numbered 1-14 (or 0-13) inclusive.
For example, starting from the unrecombined cassette, an interaction (0, 12) yields the 1-element cassette {13}
}
\details{
Calculates the generation probability distribution for cassettes
for varying values of n (step number)
}
\examples{
# Example usage:
# Assuming 'transition_probabilities' is a list of transition probabilities,
# 'steps' is the number of steps to generate, 'start_index' is the size_idx from which to start,
# and 'start_id' is the id of the cassette to initialize the Markov chain from.
# Calculate the generation probabilities for the specified parameters
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/Week2.rds")
#generation_probs <- run_markov(steps, transition_probabilities, start_index, start_id)
#generation_probs
}
