% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{sample}
\alias{sample}
\alias{sample,loxcode_experiment-method}
\title{#' Set experiment name
Get loxcode_sample object by sample name}
\usage{
sample(x, s)

\S4method{sample}{loxcode_experiment}(x, s)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{sample name}
}
\value{
loxcode_sample object corresponding to s
}
\description{
#' Set experiment name
Get loxcode_sample object by sample name
}
