% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{sample}
\alias{sample}
\alias{sample,loxcode_experiment-method}
\title{#' Set experiment name
Get loxcode_sample object by sample name}
\usage{
sample(x, s)

\S4method{sample}{loxcode_experiment}(x, s)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{s}{sample name}
}
\value{
loxcode_sample object corresponding to s
}
\description{
#' Set experiment name
Get loxcode_sample object by sample name
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object and 'sample_name' is the name of the sample to retrieve
# Retrieve the loxcode_sample object corresponding to the sample named 'sample_name'
sample_object <- get_loxcode_sample_by_sample_name(x = experiment, s = "sample_name")
sample_object
}
