% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sample_table}
\alias{sample_table}
\title{Sample Table}
\usage{
sample_table(x, c)
}
\arguments{
\item{x}{loxcode_experiment object}

\item{c}{count_matrix of samples to display}
}
\value{
A table of the samples selected
}
\description{
Tabulates the barcode counts, total number of reads, maximum complexity and consensus filtered for each sample in an experiment
}
