% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{sampnames}
\alias{sampnames}
\alias{sampnames,loxcode_experiment-method}
\title{Fetch names of samples}
\usage{
sampnames(x)

\S4method{sampnames}{loxcode_experiment}(x)
}
\arguments{
\item{x}{loxcode_experiment object}
}
\value{
names (handles) of samples
}
\description{
The names returned by this function can be used to refer to each individual sample
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object
# Retrieve the names of all samples in the experiment
sample_names <- fetch_sample_names(experiment)
sample_names
}
