% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{samptable<-}
\alias{samptable<-}
\alias{samptable<-,loxcode_experiment-method}
\title{Set sample table}
\usage{
samptable(x) <- value

\S4method{samptable}{loxcode_experiment}(x) <- value
}
\arguments{
\item{x}{lox code object}

\item{value}{value of sample table}
}
\value{
sample table
}
\description{
Set sample table
}
\examples{
# Example usage:
# Assuming 'lox_data' is a loxcode object and 'new_sample_table' is the new sample table to set
# Set the sample table of the loxcode object to 'new_sample_table'
lox <- readRDS('~/Desktop/LoxCodeR2024/LoxcodeR_app/Week2.rds')
#lox_data$samptable <- new_sample_table
#lox_data
}
