% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{saturation_multi}
\alias{saturation_multi}
\alias{saturation_multi,loxcode_experiment-method}
\title{Multiple lined saturation plot}
\usage{
saturation_multi(
  loxcode_experiment,
  loxcode_samples,
  codesets,
  colorby = "dist_orig"
)

\S4method{saturation_multi}{loxcode_experiment}(
  loxcode_experiment,
  loxcode_samples,
  codesets,
  colorby = "dist_orig"
)
}
\arguments{
\item{loxcode_experiment}{loxcode_experiment object}

\item{loxcode_samples}{A matrix of the selected samples}

\item{codesets}{A matrix of the selected code sets}

\item{colorby}{colour by dist_orig or sample}
}
\value{
Saturation plot of the loxcode sample
}
\description{
Multiple lined saturation plot
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assuming loxcode_experiment, loxcode_samples, and codesets are defined
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")
saturation_multi(lox,"N712_N501","all_codes")
}
