% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{setup_count_matrix_codesets}
\alias{setup_count_matrix_codesets}
\alias{setup_count_matrix_codesets,loxcode_experiment-method}
\title{Generate count_matrix from individual samples and add standard code_sets}
\usage{
setup_count_matrix_codesets(x, ...)

\S4method{setup_count_matrix_codesets}{loxcode_experiment}(x)
}
\arguments{
\item{x}{loxcode object to be updated}

\item{...}{count matrix to be set up}
}
\value{
count matrix for each sample
}
\description{
Generate count_matrix from individual samples and add standard code_sets
}
\examples{
# Example usage:
# Assuming 'lox_data' is a loxcode object and 'count_matrix_1', 'count_matrix_2', etc. are individual count matrices
# Generate count matrices from individual samples and add standard code sets
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/data-2024-05-20.rds")
updated_lox <- setup_count_matrix_codesets(lox)
updated_lox
}
