% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{setup_count_matrix_codesets}
\alias{setup_count_matrix_codesets}
\title{Generate count_matrix from individual samples and add standard code_sets}
\usage{
setup_count_matrix_codesets(x, ...)
}
\description{
Generate count_matrix from individual samples and add standard code_sets
}
