% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_experiment.R
\name{setup_metadata}
\alias{setup_metadata}
\alias{setup_metadata,loxcode_experiment-method}
\title{Generate metadata for the loxcode experiment}
\usage{
setup_metadata(x)

\S4method{setup_metadata}{loxcode_experiment}(x)
}
\arguments{
\item{x}{loxcode object to be updated}
}
\value{
metadata for loxcode experiment
}
\description{
Generate metadata for the loxcode experiment
}
\examples{
# Example usage:
# Assuming 'lox_data' is a loxcode object
# Generate metadata for the loxcode experiment
lox <- readRDS('~/Desktop/loxcodeR/LoxcodeR_app/Week2.rds')
experiment_metadata <- setup_metadata(lox)
experiment_metadata
}
