% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_reads}
\alias{simulate_reads}
\alias{simulate_reads,lox_casette-method}
\title{Simulate recombination of multiple casettes}
\usage{
simulate_reads(lox, c, n)

\S4method{simulate_reads}{lox_casette}(lox, c, n)
}
\arguments{
\item{lox}{original casette}

\item{c}{number of casettes}

\item{n}{number of recombination events per casette}
}
\value{
list of casettes after cre lox
}
\description{
Simulate recombination of multiple casettes
}
\examples{
# Load required packages
library(loxcodeR)

# Example usage
# Assuming lox is defined and has the required structure
# simulate_reads(lox, c = 3, n = 10)
}
