% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{size_plot}
\alias{size_plot}
\alias{size_plot,loxcode_experiment-method}
\title{Size Plot}
\usage{
size_plot(
  lox,
  sample,
  count_matrix = "all_samples",
  code_set = "all_codes",
  labels
)

\S4method{size_plot}{loxcode_experiment}(
  lox,
  sample,
  count_matrix = "all_samples",
  code_set = "all_codes",
  labels
)
}
\arguments{
\item{lox}{loxcode experiment object of current experiment}

\item{sample}{loxcode sample object}

\item{count_matrix}{A matrix of the selected samples}

\item{code_set}{A matrix of the selected code sets}

\item{labels}{A string containing sample name or alias}
}
\value{
a plot of the distribution of code sizes in the sample
}
\description{
Plots distribution of loxcode sizes found in sample
}
