% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{size_plot_sample}
\alias{size_plot_sample}
\title{Size Plot (v2)}
\usage{
size_plot_sample(sample)
}
\arguments{
\item{sample}{loxcode sample object}
}
\value{
a plot of the distribution of code sizes in the sample
}
\description{
Plots distribution of loxcode sizes found in sample. Works for loxcode_samples without loxcode_experiment
}
