% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sets.R
\name{summary_table}
\alias{summary_table}
\alias{summary_table,loxcode_experiment-method}
\title{Summary table (v3)}
\usage{
summary_table(lox, sample_set = "all_samples")

\S4method{summary_table}{loxcode_experiment}(lox, sample_set = "all_samples")
}
\arguments{
\item{lox}{loxcode_experiment object}

\item{sample_set}{sample set}
}
\value{
table of sample information
}
\description{
Summary table (v3)
}
\examples{
# Example usage:
# Assuming 'experiment' is a loxcode_experiment object and 'sample_set' is the sample set to summarize
# Generate a summary table for the specified sample set in the experiment
lox <- readRDS("~/Desktop/loxcodeR/LoxcodeR_app/Week2.rds")
summary_table <- summary_table(lox)
summary_table
}
