% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\name{valid}
\alias{valid}
\alias{valid,loxcode_sample-method}
\title{Access decoded cassette data, valid cassettes only}
\usage{
valid(x)

\S4method{valid}{loxcode_sample}(x)
}
\arguments{
\item{x}{loxcode_sample object}
}
\value{
valid cassettes
}
\description{
Access decoded cassette data, valid cassettes only
}
\examples{
# Example usage:
# Assuming 'sample' is a loxcode_sample object
# Access the decoded cassette data for valid cassettes only
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
valid_cassettes <- valid(lox@samples$N712_N501)
valid_cassettes
}
