% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loxcode_sample.R
\name{validate}
\alias{validate}
\alias{validate,loxcode_sample-method}
\title{Add cassette validation column to decoded cassette data}
\usage{
validate(x)

\S4method{validate}{loxcode_sample}(x)
}
\arguments{
\item{x}{loxcode cassette object}
}
\value{
Cassette with validation column
}
\description{
Add cassette validation column to decoded cassette data
}
\examples{
# Example usage:
# Assuming 'cassette_data' is a loxcode cassette object
# Add a validation column to the decoded cassette data
datards <- system.file("extdata","data-2024-05-20.rds",package="loxcodeR")
lox <- readRDS(datards)
validated_cassette <- validate(lox@samples$N712_N501)
validated_cassette
}
